+++ Scripts/ParseEmailFiles/ParseEmailFiles.py
@@ -3612,6 +3612,15 @@ def create_email_output(email_data, attached_emails):
     return res
 
 
+def is_email_data_populated(email_data):
+    # checks if email data has any item populated to it
+    if email_data:
+        for key, val in email_data.iteritems():
+            if val:
+                return True
+    return False
+
+
 def main():
     file_type = ''
     entry_id = demisto.args()['entryid']
@@ -3678,7 +3687,20 @@ def main():
                                                                  max_depth=max_depth)
                         output = create_email_output(email_data, attached_emails)
                     else:
-                        return_error("Could not extract email from file. Base64 decode did not include rfc 822 strings")
+                        try:
+                            # Try to open
+                            email_data, attached_emails = handle_eml(file_path, b64=False, file_name=file_name,
+                                                                     parse_only_headers=parse_only_headers,
+                                                                     max_depth=max_depth)
+                            is_data_populated = is_email_data_populated(email_data)
+                            if not is_data_populated:
+                                raise DemistoException("No email_data found")
+                            output = create_email_output(email_data, attached_emails)
+                        except Exception as e:
+                            demisto.debug("ParseEmailFiles failed with {}".format(str(e)))
+                            return_error("Could not extract email from file. Possible reasons for this error are:\n"
+                                         "- Base64 decode did not include rfc 822 strings.\n"
+                                         "- Email contained no Content-Type and no data.")
 
             except Exception as e:
                 return_error("Exception while trying to decode email from within base64: {}\n\nTrace:\n{}"