+++ Scripts/SetIfEmpty/CHANGELOG.md
@@ -1,4 +1,11 @@
 ## [Unreleased]
+
+
+## [20.1.2] - 2020-01-22
+- Add support to unicode default value
+- Consider "None" string as empty
+
+## [20.1.0] - 2020-01-07
 Fixed an issue where the transformer would fail when applied to a number field.
 
 ## [19.11.1] - 2019-11-26